seprate stakesteps component
