+++ src/components/modals/Stake/Stake.tsx
@@ -1,12 +1,10 @@
 import React, { FC, useEffect, useState } from 'react';
 import {
-	brandColors,
 	Button,
 	H4,
 	neutralColors,
 	OulineButton,
 	P,
-	SublineBold,
 } from '@giveth/ui-design-system';
 import styled from 'styled-components';
 import { BigNumber } from 'ethers';
@@ -33,6 +31,7 @@ import { StakeState } from '@/lib/staking';
 import ToggleSwitch from '../../styled-components/Switch';
 import { abi as ERC20_ABI } from '@/artifacts/ERC20.json';
 import { IModal } from '@/types/common';
+import StakeSteps from './StakeSteps';
 import type { PoolStakingConfig, RegenStreamConfig } from '@/types/config';
 
 interface IStakeModalProps extends IModal {
@@ -238,47 +237,7 @@ export const StakeModal: FC<IStakeModalProps> = ({
 								</StakeModalTitleText>
 							</StakeModalTitle>
 							<InnerModal>
-								{stakeState === StakeState.APPROVE ||
-								stakeState === StakeState.APPROVING ||
-								stakeState === StakeState.WRAP ||
-								stakeState === StakeState.WRAPPING ? (
-									<StakeStepsContainer>
-										<StakeStep>
-											<StakeStepTitle>
-												Approve
-											</StakeStepTitle>
-											<StakeStepNumber>1</StakeStepNumber>
-										</StakeStep>
-										<StakeStep>
-											<StakeStepTitle
-												disable={
-													!(
-														stakeState ===
-															StakeState.WRAP ||
-														stakeState ===
-															StakeState.WRAPPING
-													)
-												}
-											>
-												Stake
-											</StakeStepTitle>
-											<StakeStepNumber
-												disable={
-													!(
-														stakeState ===
-															StakeState.WRAP ||
-														stakeState ===
-															StakeState.WRAPPING
-													)
-												}
-											>
-												2
-											</StakeStepNumber>
-										</StakeStep>
-									</StakeStepsContainer>
-								) : (
-									<StakeStepsPlaceholder />
-								)}
+								<StakeSteps stakeState={stakeState} />
 								<AmountInput
 									setAmount={setAmount}
 									maxAmount={maxAmount}
@@ -425,60 +384,6 @@ export const StakeModal: FC<IStakeModalProps> = ({
 // 	)
 // }
 
-const StakeStepsContainer = styled(Flex)`
-	position: relative;
-	justify-content: space-evenly;
-	&::before {
-		content: '';
-		position: absolute;
-		width: 100%;
-		height: 1px;
-		border-top: 1px solid ${brandColors.giv[500]};
-		bottom: 11px;
-		z-index: 0;
-	}
-	&::after {
-		content: '';
-		position: absolute;
-		height: 1px;
-		border-top: 1px dashed ${brandColors.giv[500]};
-		left: -24px;
-		right: -24px;
-		bottom: 11px;
-		z-index: 0;
-	}
-	margin-bottom: 16px;
-`;
-
-const StakeStep = styled(Flex)`
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	width: 61px;
-	position: relative;
-	z-index: 1;
-`;
-
-interface IStakeStepState {
-	disable?: boolean;
-}
-
-const StakeStepTitle = styled(P)<IStakeStepState>`
-	margin-bottom: 8px;
-	color: ${props =>
-		props.disable ? brandColors.giv[300] : brandColors.giv['000']};
-`;
-const StakeStepNumber = styled(SublineBold)<IStakeStepState>`
-	color: ${props =>
-		props.disable ? brandColors.giv[300] : brandColors.giv['000']};
-	background-color: ${brandColors.giv[500]};
-	border: 3px solid
-		${props =>
-			props.disable ? brandColors.giv[300] : brandColors.giv['000']};
-	border-radius: 18px;
-	width: 24px;
-`;
-
 export const StakeModalContainer = styled.div`
 	width: 370px;
 	padding-bottom: 24px;
@@ -491,10 +396,6 @@ const StakeModalTitle = styled(Flex)`
 const StakeModalTitleText = styled(H4)`
 	margin-left: 54px;
 	color: ${neutralColors.gray[100]};
-`;
-
-const StakeStepsPlaceholder = styled.div`
-	padding: 13px;
 `;
 
 const InnerModal = styled.div`
